ID to Long type

+++ w-gory-java/src/main/java/com/wGory/controller/controller/WycieczkaController.java
@@ -26,7 +26,7 @@ public class WycieczkaController {
             dataType = "Integer",
             example = "1")
     @GetMapping("/{id}")
-    public ResponseEntity<Wycieczka> getWycieczkaById(@PathVariable(value = "id") Integer id) {
+    public ResponseEntity<Wycieczka> getWycieczkaById(@PathVariable(value = "id") Long id) {
         return ResponseEntity.ok(wycieczkaService.getWycieczkaById(id));
     }
 
@@ -35,10 +35,10 @@ public class WycieczkaController {
             name = "id",
             value = "Unikatowy numer identyfikujący wycieczkę w bazie danych.",
             required = true,
-            dataType = "Integer",
+            dataType = "Long",
             example = "1")
     @GetMapping("/done/{id}")
-    public ResponseEntity<Wycieczka> setWycieczkaDone(@PathVariable(value = "id") Integer id) {
+    public ResponseEntity<Wycieczka> setWycieczkaDone(@PathVariable(value = "id") Long id) {
         return ResponseEntity.ok(wycieczkaService.setWycieczkaDone(id));
     }
 }
