+++ w-gory-java/src/main/java/com/wGory/controller/repository/WycieczkaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WycieczkaRepository extends JpaRepository<Wycieczka, Integer> {
+public interface WycieczkaRepository extends JpaRepository<Wycieczka, Long> {
 }
